+++ app/src/main/java/com/example/appraisal/UI/trial/BinomialActivity.java
@@ -1,16 +1,13 @@
-package com.example.appraisal.UI;
+package com.example.appraisal.UI.trial;
 
 import android.os.Bundle;
-import android.view.LayoutInflater;
 import android.view.View;
-import android.widget.Button;
-import android.widget.EditText;
 import android.widget.TextView;
 
 import androidx.appcompat.app.AppCompatActivity;
 
 import com.example.appraisal.R;
-import com.example.appraisal.model.BinomialModel;
+import com.example.appraisal.model.trial.BinomialModel;
 
 
 public class BinomialActivity extends AppCompatActivity {
