+++ app/src/main/java/com/example/appraisal/UI/trial/NonNegIntCountActivity.java
@@ -1,12 +1,11 @@
-package com.example.appraisal.UI;
+package com.example.appraisal.UI.trial;
 
 import android.os.Bundle;
-import android.util.Log;
 import android.view.View;
 import android.widget.EditText;
 import androidx.appcompat.app.AppCompatActivity;
 import com.example.appraisal.R;
-import com.example.appraisal.model.NonNegIntCountModel;
+import com.example.appraisal.model.trial.NonNegIntCountModel;
 
 public class NonNegIntCountActivity extends AppCompatActivity {
