--- app/src/main/java/com/example/appraisal/backend/Experiment.java
@@ -1,4 +1,0 @@
-package com.example.appraisal.backend;
-
-public class Experiment {
-}
