--- app/src/main/java/com/example/appraisal/backend/Trial.java
@@ -1,4 +1,0 @@
-package com.example.appraisal.backend;
-
-abstract class Trial {
-}
