+++ app/src/main/java/com/example/appraisal/backend/trial/MeasurementTrial.java
@@ -1,0 +1,6 @@
+package com.example.appraisal.backend.trial;
+
+import com.example.appraisal.backend.trial.Trial;
+
+public class MeasurementTrial extends Trial {
+}
