+++ app/src/main/java/com/example/appraisal/backend/trial/Trial.java
@@ -1,0 +1,4 @@
+package com.example.appraisal.backend.trial;
+
+abstract class Trial {
+}
